Rename aux to auxiliary
