+++ tests/auxiliary/base_hartmann6.json
