+++ tests/auxiliary/base_hartmann6_discrete.json
