+++ tests/auxiliary/base_hartmann6_h.json
