+++ tests/auxiliary/branin4_scenario_categorical.json
