+++ tests/auxiliary/branin4_scenario_discrete.json
