+++ tests/auxiliary/chakong_haimes_scenario.json
