+++ tests/auxiliary/hartmann6_scenario_gp.json
