+++ tests/auxiliary/rs_cot_1024_scenario.json
